added enums as string; removed incorrect seeding DB; others small fixes.

+++ WebCalculator/Models/HistoryItem.cs
@@ -5,7 +5,7 @@ namespace WebCalculator.Models
     public class HistoryItem
     {
         public long Id { get; set; }
-        public double Result { get; set; }
+        public decimal Result { get; set; }
         public decimal Left { get; set; }
         public decimal Right { get; set; }
         public OperationType Type { get; set; }
